requirements.txt
amqp==5.1.1
asgiref==3.6.0
psycopg2
attrs==23.1.0
billiard==3.6.4.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.2
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-filter==23.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-timezone-field==4.2.3
djangorestframework==3.12
djangorestframework-simplejwt==5.2.2
flake8==3.9.2
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
mccabe==0.6.1
Pillow==9.4.0
prompt-toolkit==3.0.38
pycodestyle==2.7.0
pyflakes==2.3.1
drf-yasg==1.20.0
PyJWT==1.7.1
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==3.5.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.6
uwsgi
django-js-asset==2.0.0
django-mptt==0.14.0
django-multifilefield==1.0